Implement using design patterns
