+++ src/app/events/[id]/page.js
@@ -1,45 +1,11 @@
-import fs from 'fs';
+import IdContainer from "./IdContainer";
 import path from 'path';
-import Link from 'next/link';
+import fs from 'fs';
 
-
-export default async function EventDetail({ params = {}}) {
-  const { id } = await params;
-
-  const filePath = path.join(process.cwd(), 'src/app/data/events.json');
-  const fileData = fs.readFileSync(filePath, 'utf-8');
-  const events = JSON.parse(fileData);
-
-  const event = events.find(e => e.id.toString() === id);
-
-  if (!event) {
-    return <p>Event not found.</p>;
-  }
-
-  return (
-    <div>
-      <h2>{event.name}</h2>
-      <p>{event.details}</p>
-      {event.date && <p>Date and Time : {event.date} {event.time}</p>}
-      {event.location && <p>Location : {event.location}</p>}
-      <div>{event.image && (
-        <img
-          src={event.image}
-          alt={event.name}
-          style={{ maxWidth: '100%', height: 'auto' }}
-        />
-      )}</div>
-      <Link href={`/book?eventId=${event.id}`}>
-        <button className="button-link" style={{marginRight:'10px' }}>Book this Event</button>
-      </Link>
-      <Link href="/">
-        <button className="button-link">Back to Home</button>
-      </Link>
-    </div>
-  );
+export default function EventDetail({ params }) {
+  const { id } =params;
+  return <IdContainer id={id} />;
 }
-
-
 
 export async function generateStaticParams() {
   const filePath = path.join(process.cwd(), 'src/app/data/events.json');
